LibGDX, here I am.

+++ src/net/torvald/terrarum/gameactors/ai/ActorAI.kt
@@ -4,5 +4,5 @@ package net.torvald.terrarum.gameactors.ai
  * Created by minjaesong on 16-03-02.
  */
 interface ActorAI {
-    fun update(delta: Int)
+    fun update(delta: Float)
 }
